% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_calculations.R
\name{analytic_calculation_fdr}
\alias{analytic_calculation_fdr}
\title{Analytic Power Calculation for Benjamini-Hochberg FDR}
\usage{
analytic_calculation_fdr(
  associations,
  var_y = 1,
  sample_sizes = c(50, 100, 250, 500),
  n_tests = nrow(associations),
  q = 0.05,
  by = NULL,
  sample_values = FALSE,
  sample_prop = 0,
  verbose = TRUE
)
}
\arguments{
\item{associations}{A dataset of association data. It must contain the
columns described below (\code{pid}, \code{snp}, \code{beta}, \code{maf})}

\item{var_y}{The variance of the response}

\item{sample_sizes}{A set of sample sizes to test}

\item{n_tests}{The number of tests in the full dataset}

\item{q}{FDR threshold}

\item{by}{A set of variable names to aggregate power over}

\item{sample_values}{Should omitted variants (\code{n_tests - nrow(associations)})
be sampled from \code{associations}}

\item{sample_prop}{If \code{sample_values = TRUE}, the proportion of the
sampled values that should be sampled from \code{associations}. The
remaining proportion of sampled values (\code{1 - sample_prop}) will be null
values.}

\item{verbose}{If \code{TRUE}, messages will be output during the course of
the function.}
}
\value{
The estimated statistical power for each gene-SNP association at an
FDR of \code{q}.
}
\description{
This function calculates the long-run statistical power to detect eQTLs when
using the Benjamini-Hochberg procedure to determine significance.
}
\examples{
data("associations")
bh_results.dt <- analytic_calculation_fdr(associations.dt)
}
