% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_calculations.R
\name{analytic_calculation_fixed}
\alias{analytic_calculation_fixed}
\title{Analytic Power Calculation for a Fixed p-value Threshold}
\usage{
analytic_calculation_fixed(
  associations,
  var_y = 1,
  sample_sizes = c(50, 100, 250, 500),
  n_tests = nrow(associations),
  pval_threshold = 0.05/n_tests,
  verbose = TRUE,
  by = NULL
)
}
\arguments{
\item{associations}{A dataset of association data. It must contain the
columns described below (\code{pid}, \code{snp}, \code{beta}, \code{maf})}

\item{var_y}{The variance of the response}

\item{sample_sizes}{A set of sample sizes to test}

\item{n_tests}{The number of tests in the full dataset}

\item{pval_threshold}{The level the FWER is controlled at by the Bonferroni
adjustment.}

\item{verbose}{If \code{TRUE}, messages will be output during the course of
the function.}

\item{by}{A set of variable names to aggregate power over}
}
\value{
The estimated statistical power for each gene-SNP association when
the \eqn{p}-value threshold is set at \code{pval_threshold}.
}
\description{
This function calculates the long-run statistical power to detect eQTLs when
using a fixed p-value threshold or the Bonferroni method to determine
significance.
}
\examples{
data("associations")
bonf_results.dt <- analytic_calculation_fixed(associations.dt)
}
