% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{associations.dt}
\alias{associations.dt}
\title{Example Gene-SNP Associations}
\format{
A data.table with 1000 rows and 6 variables:
\describe{
  \item{pid}{gene ID}
  \item{snp}{SNP ID}
  \item{qtl_tag}{whether the SNP is considered a true eQTL}
  \item{beta}{true marginal effect between the SNP and the gene}
  \item{maf}{minor allele frequency of SNP}
  \item{cis_flag}{whether the SNP is within 1 megabases of the gene}
}
}
\usage{
data("associations")
}
\description{
This dataset is an example of required data for the \code{associations}
argument in the \code{\link{generate_coefficients}} and analytic calculation
functions.
}
