% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{eqtlpower}
\alias{eqtlpower}
\title{eqtlpower: A package for estimating eQTL Study Power}
\description{
The eqtlpower package provides two methods to estimate statistical power
for eQTL studies: semi-analytic and analytic.
}
\section{Analytic Power Calculation}{

The eqtlpower package provides two methods for calculating statistical power analytically depending on the multiple comparisons adjustment required. The \code{\link{analytic_calculation_fdr}} function provides analytic calculation for the Benjamini-Hochberg procedure and the \code{\link{analytic_calculation_fixed}} function can be used for either fixed p-value thresholds (e.g. p < 0.05) or the Bonferroni adjustment.
}

\section{Semi-Analytic Power Calculation}{

The semi-analytic power calculation method generates realisations of SNP-gene marginal coefficients based on the linkage disequilibrium that exists between variants. This method requires the use of two functions sequentially. Firstly, the user can generate a number of realisations of the coefficients using the \code{\link{generate_coefficients}} function; this results in a dataset containing a column for each realisation of the coefficients. This dataset can either be analysed by the user with custom code for a novel methodology, or the \code{\link{analyse_coef_results}} function can be used to calculate the estimated statistical power based on the multiple testing methods contained in \code{\link[stats]{p.adjust}}. Note there is an option to choose whether to calculate the power for either variants or genes in the analysis function.
}

\section{Visualising the Estimated Power}{

The estimated statistical power for a number of sample sizes can be visualised using the \code{\link{plot_power_results}} function. This provides a basic visualisation of the results as a ggplot2 object which can be used with any of the usual ggplot2 layers. Note that this is just a basic method of visualising results; advanced users can easily produce their own custom plots using the power results datasets directly.
}

