% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_calculations.R
\name{fdr_p_threshold}
\alias{fdr_p_threshold}
\title{Calculate Whether \eqn{p}-value Threshold Results in a Given FDR}
\usage{
fdr_p_threshold(pval_threshold, M = length(ncp), q = 0.05, ncp, pi0 = 1)
}
\arguments{
\item{pval_threshold}{\eqn{p}-value threshold to check against}

\item{M}{Number of comparisons being made}

\item{q}{Desired level of FDR}

\item{ncp}{Non-centrality parameter}

\item{pi0}{\eqn{pi_0} parameter for pFDR}
}
\value{
The value of the equation given in the description.
}
\description{
This function calculates the FDR for a given set of \eqn{p}-values if the
threshold is set at \code{pval_threshold}. It achieves this by solving
\eqn{(M \pi_0 \times pval_threshold)/(\sum(power for each eQTL))}. This
function is used within the \code{\link{analytic_calculation_fdr}} function.
}
\examples{
data("associations")
test_ncp <- 515 * 2 * associations.dt$maf * (1 - associations.dt$maf) * associations.dt$beta^2
fdr_p_threshold(0.05, ncp = test_ncp)
}
