% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coefficients.R
\name{generate_coefficients}
\alias{generate_coefficients}
\title{Generate eQTL coefficients based on Summary Statistics and Linkage
Disequilibrium}
\usage{
generate_coefficients(
  associations,
  ld,
  var_y = 1,
  sample_sizes = c(50, 100, 250, 500),
  reps = 100,
  verbose = TRUE,
  seed = 12345
)
}
\arguments{
\item{associations}{A dataset of associations to simulate.}

\item{ld}{A linkage disequilibrium dataset generated by PLINK.}

\item{var_y}{The variance of the response}

\item{sample_sizes}{A vector of sample sizes to test.}

\item{reps}{Number of realisations to generate.}

\item{verbose}{If \code{TRUE}, messages will be output during the course of
the function.}

\item{seed}{The random seed that will be set before generating coefficients}
}
\value{
A dataset containing metadata for each gene-SNP association with
\code{reps} columns where each is a realisation.
}
\description{
This function generates a series of realisations of the SNP-gene association
coefficients based on the true association value (\code{beta}) and the
linkage disequilibrium that exists between variants.
}
\examples{
\donttest{
data("associations")
data("ld")

sa_coefs <- generate_coefficients(associations.dt, ld.dt)
}
}
