% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ld.dt}
\alias{ld.dt}
\title{Example Estimated Linkage Disequilibrium}
\format{
A data.table with 1064 rows and 9 variables:
\describe{
  \item{CHR_A}{chromosome on which SNP A is located}
  \item{BP_A}{location on the chromosome in base pairs where SNP A is located}
  \item{SNP_A}{ID of SNP A}
  \item{MAF_A}{minor allele frequency of SNP A}
  \item{CHR_B}{chromosome on which SNP B is located}
  \item{BP_B}{location on the chromosome in base pairs where SNP B is located}
  \item{SNP_B}{ID of SNP B}
  \item{MAF_B}{minor allele frequency of SNP B}
  \item{R}{estimated linkage disequilibrium between SNP A and SNP B}
}
}
\usage{
data("ld")
}
\description{
This dataset is an example of required data for the \code{ld}
argument in the \code{\link{generate_coefficients}}. Each row of this dataset
is a SNP pair and the estimated linkage disequilibrium between them. This
dataset was produced by the PLINK program.
}
