% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_power_results}
\alias{plot_power_results}
\title{Plot the Estimated eQTL Power Results}
\usage{
plot_power_results(
  power.dt,
  by = NULL,
  title = "Estimated Statistical Power for eQTL Study"
)
}
\arguments{
\item{power.dt}{Result dataset from either \code{analyse_coef_results} or \code{analytic_calculation_fdr}.}

\item{by}{A variable to colour lines by.}

\item{title}{Plot title}
}
\value{
A ggplot of the power curves
}
\description{
Generate a simple visualisation of the estimated statistical power from
either the \code{\link{analyse_coef_results}} or the analytic functions.
}
\examples{
\donttest{
data("associations")
data("ld")

sa_coefs <- generate_coefficients(associations.dt, ld.dt)
sa_results <- analyse_coef_results(associations.dt, sa_coefs)

plot_power_results(sa_results)
}

bh_results.dt <- analytic_calculation_fdr(
  associations.dt[qtl_tag == TRUE],
  by = c("cis_flag")
)

plot_power_results(bh_results.dt, by = "cis_flag")
}
