% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_coefficients.R
\name{sa_power_calc_genes}
\alias{sa_power_calc_genes}
\title{Calculate statistical power to detect eGenes based on adjusted p-values}
\usage{
sa_power_calc_genes(
  coef_results,
  adjusted_matrix,
  method_name,
  threshold = 0.05,
  interval = 0.95,
  by = NULL
)
}
\arguments{
\item{coef_results}{A set of results for a given sample size from the
\code{generate_coefficients} function}

\item{adjusted_matrix}{A matrix of adjusted p-values}

\item{method_name}{A label for the method used to create
\code{adjusted_matrix}}

\item{threshold}{A threshold for values in \code{adjusted_matrix} to be
considered significant}

\item{interval}{Confidence interval level}

\item{by}{A set of variable names to aggregate power over}
}
\value{
The mean statistical power to detect eGenes over the realisations for each group according to the \code{by} variables.
}
\description{
This function takes a set of coefficient results from
\code{\link{generate_coefficients}} and a adjusted p-value matrix generated
based on these values and calculates how many eGenes were detected. Note this
function is not intended for typical usage as it is used within the
 \code{\link{analyse_coef_results}} function.
}
